Update properties.dart
If the string is too long (only short sentences are possible right now) the string is truncated, so not all text is displayed. 

With this proposed change the following will happen: if the value.toString() is longer than the available space, the text will be horizontally scrollable within the Container. If the text fits within the available space, it will be displayed normally without any scrolling.